% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagerConfirm.R
\name{ChecknoteConfirm}
\alias{ChecknoteConfirm}
\title{整单确认}
\usage{
ChecknoteConfirm(statementNo)
}
\arguments{
\item{statementNo}{第二个参数}
}
\value{
两个数的和
}
\description{
整单确认
}
\examples{
ChecknoteConfirm
}
