% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagerConfirm.R
\name{ChecknotedeleteMessage}
\alias{ChecknotedeleteMessage}
\title{对账单删除}
\usage{
ChecknotedeleteMessage(statementID)
}
\arguments{
\item{statementID}{}
}
\value{
两个数的和
}
\description{
对账单删除
}
\examples{
ChecknotedeleteMessage
}
