% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagerConfirm.R
\name{ChecknoteInvoice}
\alias{ChecknoteInvoice}
\title{开票申请}
\usage{
ChecknoteInvoice(statementNo)
}
\arguments{
\item{statementNo}{第二个参数}
}
\value{
两个数的和
}
\description{
开票申请
}
\examples{
ChecknoteInvoice
}
